updated soulmeter no longer provide vision in filename fixed permission problem cased by WinDivert.dll WinDivert64.sys

+++ soulmeter_handler.py
@@ -1,45 +1,55 @@
-#Code by AkinoAlice@Tyrant_Rex
+# Code by AkinoAlice@Tyrant_Rex
 
-import requests, os, shutil, re, json
+import requests
+import os
+import re
+import json
 
-from webbrowser import open_new
 from bs4 import BeautifulSoup
+from webbrowser import open_new
+from zipfile import ZipFile
+
 
 class PermissionError(Exception):
     pass
 
+
 class handler:
     def __init__(self) -> None:
-        self.StartSoulWorker :bool
-        self.KeepArchive :bool
-        self.RepoOwner :str
-        self.vision :str
+        self.StartSoulWorker: bool
+        self.KeepArchive: bool
+        self.RepoOwner: str
+        self.vision: str
 
-        try:
-            with open("./config.json","r",encoding="utf-8") as f:
-                setting = json.load(f)
-
-        except FileNotFoundError:
+        if not os.path.exists(f"./config.json"):
             setting = {
                 "StartSoulWorker": True,
                 "KeepArchive": False,
+                "KeepFonts": False,
                 "RepoOwner": "neonr-0",
             }
             with open("./config.json", "w") as setting_file:
-                setting_file.write(json.dumps(setting,indent=4))
+                setting_file.write(json.dumps(setting, indent=4))
+
+        with open("./config.json", "r", encoding="utf-8") as f:
+            setting = json.load(f)
 
         self.StartSoulWorker = setting["StartSoulWorker"]
         self.KeepArchive = setting["KeepArchive"]
+        self.KeepFont = setting["KeepFonts"]
         self.RepoOwner = setting["RepoOwner"]
         self.vision = ""
 
-    #check update
+    # check update
     def check_vision(self) -> None:
-        soup = BeautifulSoup(requests.get(f"https://github.com/{self.RepoOwner}/SoulMeter").text, "lxml")
-        soup = soup.find(href=re.compile(f"/{self.RepoOwner}/SoulMeter/releases/tag/*"))
-        self.vision = soup["href"].replace(f"/{self.RepoOwner}/SoulMeter/releases/tag/","")
+        soup = BeautifulSoup(requests.get(
+            f"https://github.com/{self.RepoOwner}/SoulMeter").text, "lxml")
+        soup = soup.find(href=re.compile(
+            f"/{self.RepoOwner}/SoulMeter/releases/tag/*"))
+        self.vision = soup["href"].replace(
+            f"/{self.RepoOwner}/SoulMeter/releases/tag/", "")
 
-    #update soulmeter
+    # update soulmeter
     def update(self) -> None:
         for soulmeter_file in os.listdir("./soulmeter"):
             if soulmeter_file.endswith(".exe"):
@@ -48,14 +58,26 @@ class handler:
                 os.remove(f"./soulmeter/{soulmeter_file}")
 
         with open(f"./archive/{self.vision}.zip", "wb") as f:
-            f.write(requests.get(f"https://github.com/{self.RepoOwner}/SoulMeter/releases/download/{self.vision}/SoulMeter-v{self.vision}.zip").content)
+            f.write(requests.get(
+                f"https://github.com/{self.RepoOwner}/SoulMeter/releases/download/{self.vision}/SoulMeter-v{self.vision}.zip").content)
 
-        shutil.unpack_archive(f"./archive/{self.vision}.zip", "./soulmeter")
+        with ZipFile(f"./archive/{self.vision}.zip", "r") as zip_file:
+            for file_name in zip_file.namelist():
+                if file_name in ["WinDivert.dll", "WinDivert64.sys"]:
+                    continue
+
+                if ".ttf" in file_name and self.KeepFont and os.path.exists("./soulmeter/Font"):
+                    continue
+
+                zip_file.extract(file_name, r"./soulmeter")
+
+        os.rename("./soulmeter/SoulMeter.exe",
+                  f"./soulmeter/SoulMeter {self.vision}.exe")
 
         if not self.KeepArchive:
             os.remove(f"./archive/{self.vision}.zip")
 
-    #startup
+    # startup
     def startup(self) -> None:
         if not os.path.exists("./soulmeter"):
             os.mkdir("./soulmeter")
@@ -75,12 +97,13 @@ class handler:
         self.update()
         self.start_soulworker()
 
+
 if __name__ == "__main__":
     try:
         sw = handler()
         sw.main()
-    except:
-        raise PermissionError("""
+    except Exception as e:
+        raise PermissionError(f"""
                 Please follow the following instructions:
                 1. Right click the program
                 2. Select "Properties" -> "Compatibility".
@@ -90,4 +113,6 @@ if __name__ == "__main__":
                 1.右鍵單擊程序
                 2.選擇屬性 -> 兼容性
                 3.檢查是否選中“以管理員身份運行此程序”的複選框
+                ----------------------------------------------------------------
+                {e}
         """)
